reorg and style
