+++ frontend/src/metabase/visualizer/components/DragOverlay/DragOverlay.tsx
@@ -6,7 +6,7 @@ import {
 import type { DraggedItem } from "metabase-types/store/visualizer";
 
 import { ColumnsListItem } from "../DataImporter/ColumnsList/ColumnsListItem";
-import { WellItem } from "../VisualizationCanvas/WellItem";
+import { WellItem } from "../VisualizationCanvas/wells/WellItem";
 
 interface DragOverlayProps {
   item: DraggedItem;
