+++ frontend/src/metabase/visualizer/components/VisualizationCanvas/VisualizationCanvas.tsx
@@ -30,10 +30,10 @@ import type { RawSeries } from "metabase-types/api";
 import { TabularPreviewModal } from "../TabularPreviewModal";
 import { useVisualizerUi } from "../VisualizerUiContext";
 
-import { HorizontalWell } from "./HorizontalWell";
-import { ScatterFloatingWell } from "./ScatterFloatingWell";
-import { VerticalWell } from "./VerticalWell";
 import S from "./VisualizationCanvas.module.css";
+import { HorizontalWell } from "./wells/HorizontalWell";
+import { ScatterFloatingWell } from "./wells/ScatterFloatingWell";
+import { VerticalWell } from "./wells/VerticalWell";
 
 function disableAxisLabels(rawSeries: RawSeries) {
   return produce(rawSeries, (draft) => {
