+++ frontend/src/metabase/visualizer/components/VisualizationCanvas/wells/HorizontalWell/CartesianHorizontallWell.tsx
@@ -1,4 +1,5 @@
 import { useDraggable, useDroppable } from "@dnd-kit/core";
+import cx from "classnames";
 import { useMemo } from "react";
 
 import { useDispatch, useSelector } from "metabase/lib/redux";
@@ -19,6 +20,7 @@ import { isDate, isString } from "metabase-lib/v1/types/utils/isa";
 import type { DatasetColumn } from "metabase-types/api";
 
 import { WellItem } from "../WellItem";
+import S from "../well.module.css";
 
 export function CartesianHorizontalWell({ style, ...props }: FlexProps) {
   const display = useSelector(getVisualizationType);
@@ -74,31 +76,18 @@ export function CartesianHorizontalWell({ style, ...props }: FlexProps) {
     dispatch(removeColumn({ name: dimension.name }));
   };
 
-  const borderColor = canHandleActiveItem
-    ? "var(--mb-color-brand)"
-    : "var(--border-color)";
-
   return (
     <Flex
       {...props}
-      bg={canHandleActiveItem ? "var(--mb-color-brand-light)" : "bg-light"}
-      p="sm"
-      wrap="nowrap"
-      gap="sm"
+      className={cx(S.Well, {
+        [S.isOver]: isOver,
+        [S.isActive]: canHandleActiveItem,
+      })}
       style={{
         ...style,
         height: "100%",
         overflowX: "auto",
         overflowY: "hidden",
-        borderRadius: "var(--border-radius-xl)",
-        border: `1px solid ${borderColor}`,
-        transform: canHandleActiveItem ? "scale(1.025)" : "scale(1)",
-        transition:
-          "transform 0.2s ease-in-out 0.2s, border-color 0.2s ease-in-out 0.2s, background 0.2s ease-in-out 0.2s",
-        outline:
-          isOver && canHandleActiveItem
-            ? "1px solid var(--mb-color-brand)"
-            : "none",
       }}
       ref={setNodeRef}
     >
