+++ frontend/src/metabase/visualizer/components/VisualizationCanvas/wells/ScatterFloatingWell/index.ts
