+++ frontend/src/metabase/visualizer/components/VisualizationCanvas/wells/VerticalWell/FunnelVerticalWell.tsx
