+++ frontend/src/metabase/visualizer/components/VisualizationCanvas/wells/VerticalWell/VerticalWell.tsx
