+++ frontend/src/metabase/visualizer/components/VisualizationCanvas/wells/well.module.css
@@ -1,0 +1,30 @@
+.Well {
+  flex-wrap: nowrap;
+  background-color: var(--mb-color-bg-light);
+  /* bg={canHandleActiveItem ? "var(--mb-color-brand-light)" : "bg-light"} */
+  padding: var(--mantine-spacing-sm);
+  gap: var(--mantine-spacing-sm);
+  border-radius: var(--border-radius-xl);
+  border: 1px dashed var(--border-color);
+  transform: scale(1);
+  transition:
+    transform 0.2s ease-in-out 0.2s,
+    border-color 0.2s ease-in-out 0.2s,
+    background 0.2s ease-in-out 0.2s;
+  outline: none;
+}
+
+.isActive {
+  border-color: var(--mb-color-brand);
+  background-color: var(--mb-color-brand-lighter);
+  transform: scale(1.025);
+}
+
+.isOver {
+  outline: 1px solid var(--mb-color-brand);
+  background-color: var(--mb-color-brand-light);
+}
+
+.defaultBorderRadius {
+  border-radius: var(--default-border-radius);
+}
